Consolidate hooks into one file and fix image data
